Fixed problem with batch_maker
The process works now from DEM to postprocess, completely. It fails thought when you have an image with empty pixels (nan) in the extremes and it considers the path to reach these points. When this happens it goes out of boundaries.
Best way to fix it, get binary dimensions from path max and min pixel values, then it could happen that the binary images are bigger than the rest. But it should be the same for straight DEMs.